+++ pypsdm/plots/results/connector_plot.py
@@ -1,19 +1,16 @@
-from typing import Optional, Union
+from typing import Optional
 
 import matplotlib.pyplot as plt
 import numpy as np
-from loguru import logger
 from matplotlib.axes import Axes
 
 from pypsdm.models.gwr import GridWithResults
-from pypsdm.models.input.connector.lines import Lines
 from pypsdm.models.result.grid.connector import ConnectorResult
 from pypsdm.models.result.grid.transformer import Transformer2WResult
 from pypsdm.plots.common.line_plot import ax_plot_time_series
 from pypsdm.plots.common.utils import (
     BLUE,
     FIGSIZE,
-    ax_plot_secondary_axis,
     set_subplot_title,
     set_suptitle,
     set_title,
@@ -26,7 +23,6 @@ def plot_transformer_s(
     gwr: GridWithResults,
     resolution: str,
     include_utilisation: bool = True,
-    show_util_grid_lines: bool = False,
     title: Optional[str] = None,
     fill_from_index: bool = False,
     fill_between=None,
@@ -134,86 +130,6 @@ def ax_plot_transformer_utilization(
         **kwargs,
     )
     ax.set_ylabel("Transformer Utilization")
-
-
-def plot_line_current(
-    res: ConnectorResult,
-    side: str,
-    resolution: str,
-    include_utilisation: bool = False,
-    i_max_src: Optional[Union[float, Lines]] = None,
-    title: Optional[str] = None,
-    fill_from_index: bool = False,
-    fill_between=None,
-    set_x_label=True,
-    **kwargs,
-):
-    line_name = res.name if res.name else res.input_model
-    if title is None:
-        title = f"Current Magnitude Line: {line_name}"
-    fig, ax = plt.subplots(figsize=FIGSIZE)
-    ax_plot_line_current(
-        ax,
-        res,
-        side,
-        resolution,
-        i_max_src=i_max_src,
-        include_utilisation=include_utilisation,
-        fill_from_index=fill_from_index,
-        fill_between=fill_between,
-        set_x_label=set_x_label,
-        **kwargs,
-    )
-    ax.set_ylabel("Current in Ampere")
-    set_title(ax, title)
-    return fig, ax
-
-
-def ax_plot_line_current(
-    ax: Axes,
-    res: ConnectorResult,
-    side: str,
-    resolution: str,
-    include_utilisation: bool = True,
-    i_max_src: Optional[Union[float, Lines]] = None,
-    fill_from_index: bool = False,
-    fill_between=None,
-    set_x_label=True,
-    **kwargs,
-):
-    if len(res.i_a_mag) == 0:
-        raise ValueError("Line current time series is empty. No data to plot")
-
-    current = get_connector_current(side, res)
-
-    ax = ax_plot_time_series(
-        ax,
-        current,
-        res.entity_type,
-        resolution,
-        fill_from_index=fill_from_index,
-        fill_between=fill_between,
-        set_x_label=set_x_label,
-        **kwargs,
-    )
-    ax.set_ylabel("Current Magnitude in Ampere")
-
-    if include_utilisation:
-        if i_max_src is None:
-            logger.warning("No i_max_src provided. Cannot plot utilisation")
-            return
-        if isinstance(i_max_src, float) or isinstance(i_max_src, int):
-            line_i_max = i_max_src
-        elif isinstance(i_max_src, Lines):
-            line_i_max = i_max_src.subset(res.input_model).i_max.iloc[0]
-        else:
-            raise ValueError(
-                f"i_max_src should be either float or Lines, got {type(i_max_src)}"
-            )
-        utilisation = current / line_i_max
-        ax_plot_secondary_axis(
-            ax, utilisation, "Line Utilisation", show_secondary_grid_lines=True
-        )
 
 
 def plot_connector_current(